requirements.txt
bs4
pathlib
pyyaml
requests
#tkinter
urllib3